exposing server as a library
